+++ src/components/LegacyQuoteHistoryModal.jsx
@@ -2093,6 +2093,7 @@ export function QuoteHistoryView({
             <QuoteDecisionDebtPanel
               organizationId={organizationId}
               quoteId={focusedQuote.id}
+              tenantTimeZone={tenantTimeZone}
               available={Boolean(organizationId)}
               onOpenWorkflow={onOpenWorkflow}
             />
@@ -2412,6 +2413,7 @@ export function QuoteHistoryView({
           <QuoteDecisionDebtPanel
             organizationId={organizationId}
             quoteId={focusedQuote.id}
+            tenantTimeZone={tenantTimeZone}
             available={Boolean(organizationId)}
             onOpenWorkflow={onOpenWorkflow}
           />